+++ crates/agentgateway/src/mcp/mod.rs
@@ -4,7 +4,6 @@ pub(crate) mod guardrails;
 mod handler;
 mod mergestream;
 mod rbac;
-mod resolve;
 mod router;
 mod routing;
 mod session;
