+++ examples/mcp-apps/README.md
@@ -43,4 +43,4 @@ cargo run -- -f examples/mcp-apps/config-multiplex.yaml
 
 This config serves two Apps-enabled servers behind one endpoint. By default, multiplexing prefixes tool names with their target (`map_`, `monitor_`) — but a rendered app can itself call tools, using tool names hardcoded in the app's HTML, which never carry that prefix. The system-monitor app calls `poll-system-stats` by its unprefixed name, so with prefixed names the dashboard renders but its live updates fail.
 
-The config sets `prefixMode: never` to fix this: tool names are exposed unprefixed and the gateway routes each call by discovering which target serves the name. `ui://` URIs are still rewritten to carry the target (`ui://monitor+system-monitor/mcp-app.html`) — hosts only ever see URIs the gateway produced, so that stays transparent. The trade-off: tool names must be unique across targets. Names served by more than one target are dropped from listings and calls to them fail with an error naming the colliding targets.
+The config sets `prefixMode: never` to fix this: tool names are exposed unprefixed and the gateway routes each call by discovering which target serves the name. `ui://` URIs are still rewritten to carry the target (`ui://monitor+system-monitor/mcp-app.html`) — hosts only ever see URIs the gateway produced, so that stays transparent. The trade-off: tool names must be unique across targets. Names served by more than one target are dropped from listings — regardless of the caller's authorization — and calls to them fail the same as unknown names; the colliding targets are logged at the gateway.
